Support some more expressions.
Add support for variable set, static get and set, and property get and set.

There is some duplicated code that needs some abstraction.  The expression
temporary is unused, we use the "let" pattern consistently.  This might
result in worse code in the unoptimized JIT code but is the same code after
conversion to SSA.