[flutter-aot] Emit slightly simpler code for the `let` pattern.
Instead of pushing the value of the let variable and then dropping the let
variable from underneath (sliding the value down), simply use the value of
the let variable as a value on the expression stack.

The SSA graph is actually the same, but the unoptimized JIT code might be
simpler so we should consider pushing this to the upstream dart-sdk
repository.