[flutter-aot] Assign indexes to temporary locals in a second pass.
Variable allocation is done after flow graph construction, so we can't
really use the number of stack-allocated variables in a method until after
that.  This is used to compute the fp-relative index of temporary locals
that are allocated in the expression stack.

Set these indexes to be relative to the base of the expression stack and
adjust them after variable allocation by a loop over the list of temporary
variables.