[flutter-aot] Support for while and do loops.
Loops get a stack overflow check in the body.  We don't return a fragment
that starts with a block entry, either JoinEntry or TargetEntry, because we
cannot append them to another fragment without peeking at the first
instruction and adding a Goto.

Do loops have an 'empty' block as the true branch target of the condition
because the flow graph language statically requires us to split critical
edges (the target of a branch can't be a JoinEntry).