[flutter-aot] Huge refactoring of the flow graph builder.
Add helper functions to build Flow Graph instructions.  They return
fragments because then they can easily return multiple or zero instructions.
They implicitly find inputs on the operand stack.  For instructions whose
value can be used they leave the value on the operand stack as a side
effect.

Pushed arguments are contained in the operand stack.  Since the top of the
operand stack is used to assign indexes to operands when they are pushed and
since the pushed arguments should not be counted, they are assigned dummy
indexes that will give the correct index to operands pushed on top of them.
These indexes are cleared when they are removed from the operand stack
because PushArgument instructions are always unused.

Temporary local variables backed by the expression stack need to ensure that
the are assigned a temp index or else they will not necessarily be
materialized.  The SSA temp index field, which is otherwise unused at this
point in the pipeline, is used as a flag to indicate that they are used and
this flag is cleared when they are removed from the operand stack.

R=asgerf@google.com, kustermann@google.com

Review URL: https://chromereviews.googleplex.com/428737013 .