[flutter-aot] Support contexts in for loops and catch bodies.
For loops need to copy the context before the update expression if any
of the variables declared in the initializers are captured.  There is
a flow graph instruction that does this.

Try/catch records the context on entry to the try and then restores it
on entry to the catch.  We use a function-local variable per catch
nesting level for this, which is parallel to the variables used for
the exception and stack trace.

R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/444787013 .