[flutter-aot] Support Dil -> CoreLibrary forwarding constructor calls
Calls to forwarding constructors (e.g. `new ReceivePort()`) get translated
to a `StaticInvocation` where the target is a static method which calls the
real constructor.

But if we use the VMs core libraries, we need to account for that when resolving
forwarding factory constructors.

This change recognizes if a `StaticInvocation` results in a real constructor
target, and if so, treats it as a normal constructor call.

R=kmillikin@google.com

Review URL: https://chromereviews.googleplex.com/444197013 .