[flutter-aot] Make dart::Function objects factories instead of static methods if they take type parameters
This is necessary because a factory constructor (although static) can have type
arguments.

This mimics the current implementation in the DartVM with one exception: If a
factory constructor does not have type parameters, we will treat it as a normal
static method (whereas the DartVM treats them as factories nevertheless and
therefore passes it a `null` TypeArguments vector).

R=vegorov@google.com

Review URL: https://chromereviews.googleplex.com/451017013 .