[flutter-aot] Make DartTypeTranslator reusable, add ActiveClass/ActiveClassScope
The ActiveClass/ActiveClassScope objects make it easy to record which class
we're currently acting on. This will mainly be used to resolve generic type
parameters.

The DartTypeTranslator becomes (exactly as ConstantEvaluator) reusable. This
means that a FlowGraphBuilder needs exactly one DartTypeTranslator (and one
ConstantEvaluator). It will know what ActiveClass we're currently dealing with
and can therefore resolve references to type parameters.

This is preparation work for generic support.

R=vegorov@google.com

Review URL: https://chromereviews.googleplex.com/453747013 .