Emit unwinding directives when building assembly snapshot.
This allows us to get call-graph profiles from the actual device via simpleperf.

We emit both DWARF and ARM unwinding directives just for the sake of completeness.

One unsolved issue is unwinding through DartInvokeCode stub. It seems that
libunwind is unable to unwind through it because we are restoring FP but
not restoring SP - and the caller of DartInvokeStub has a SP based unwind
record.

I attempted to specify DW_CFA_val_offset rule for SP restoration but that did not help much, so for now we will have to accept that those stacks are missing from the collected profile.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/491837013 .