[flutter-aot] Add a build step to concatenate core library patches.
There is a tool in dart-lang/dev_compiler that applies patch files to the
SDK's libraries.  The tool expects a single patch file per library in a
directory named `patch`.  The simplest way to do this with the VM's library
sources is to simply concatenate them.

When a VM binary is built (including dart_precompiled_runtime, gen_snapshot,
etc.) generate the concatenated patch files.  They are put in the build
directory `out/<ModeArch>/obj/gen/patch`.

This change does not include changes required to the dev_compiler patch_sdk
tool, changes required to the analyzer, and changes required to the VM's
patch files to allow them to be resolved by the analyzer.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/488927014 .