[flutter-aot] Copy Dart dev_compiler's SDK patching scripts.
Add a copy of DDC's library patching scripts to the tools directory.
They do not work with the VM's patched SDK because they are customized
for dart2js libraries and use some hard-coded paths.

I will modify them to also work with the VM's patched SDK as a separate
change, so the DDC team can see the diff.  The long-term plan is that
DDC will also use this script from the dart-lang/sdk repository.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/488237013 .