[flutter-aot] Modify the SDK patching scripts to work for the VM.
The scripts now take a 'mode' option which can be 'vm' or 'ddc'.  For
'vm' they patch the Dart SDK using patches generated by a debug build of
the VM.  'ddc' doesn't work.

There are a couple of other changes needed: the analyzer needs to be
able to find all the VM's bootstrap libraries and it needs to allow
const factory constructors with a native body.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/488247013 .