Replace static const fields in ClassID with static getters.
After bootstrapping from Dart sources, the VM installs 70 static const
fields on class ClassID in library dart:_internal.  Seven of these are used
in the core library implementation, the rest seem to be unused.

The analyzer cannot resolve these fields because it doesn't see them in Dart
code.  Instead, make them static getters in the implementation of class
ClassID.

The static const fields compiled to their value.  In order to have no
performance regression for unoptimized code, the compiler needs to recognize
native static getter calls and inline them.  The only native static getters
are only the ones introduced in this change and existing ones that compile
to a single NativeCall instruction.  (We could do the same for other native
static calls, but they may be a lot of code and dealing with the arguments
in the flow graph is not simple.)

The inlined NativeCall instructions are simpler to handle if NativeCall in
the flow graph IR does not require an AST node.  This is a change we will
need to compile the core libraries from the Kernel binary anyway.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/506777013 .