Use the expected parameter indexing scheme for intrinsics.
Intrinsics that are compiled from a flow graph use a different indexing
scheme from other Dart code.  They are indexed right-to-left beginning with
0 or 1 depending on the platform, instead of left-to-right beginning with 0.

If we do not use the indexing scheme we will propagate the wrong parameter
types and possibly index the parameters out of bounds.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/504247013 .