[flutter-aot] Rewrite segv continuation asm code to be a [Stub]
Unfortunately it is not possible to use an [Assembler] and just generate a `.S`
or `.o` file and link it together with the runtime.

This CL uses therefore dynamic linking: Both, the address of the stub
entry and the address of the `InterruptContinuation` function, will be available
via the [Thread] object.

Additionally the CL blocks SIGPROF during the SEGV signal handler.

R=kmillikin@google.com

Review URL: https://chromereviews.googleplex.com/505947014 .

NOTE: This is not a regular [Stub]: There is no caller, we therefore also don't have the [Stub] code in a register. That's why I'm not using some of the helper methods for stub entry/exit.