Support context-allocated catch parameters in redundancy elimination.
The code in TryCatchAnalyzer was written assuming that the hidden exception
and stack trace variables used for implementing rethrow were always
stack-allocated.  This isn't necessarily the case, which can be seen from
the code for CatchBlockEntryInstr::EmitNativeCode (x64, for example):

  // Initialize exception and stack trace variables.
  if (exception_var().is_captured()) {
    ASSERT(stacktrace_var().is_captured());
    __ StoreIntoObject(
        CTX,
        FieldAddress(CTX, Context::variable_offset(exception_var().index())),
        kExceptionObjectReg);
    __ StoreIntoObject(
        CTX,
        FieldAddress(CTX, Context::variable_offset(stacktrace_var().index())),
        kStackTraceObjectReg);
  } else {
    __ movq(Address(RBP, exception_var().index() * kWordSize),
            kExceptionObjectReg);
    __ movq(Address(RBP, stacktrace_var().index() * kWordSize),
            kStackTraceObjectReg);
  }

Don't assume that.  Also, don't assume that a block expression cannot
contain exits such as throw or rethrow, because there is nothing that
prohibits it.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/513687014 .