Roll Kernel to fix an async bug.
There was a bug in the translation of await in finally when there
was a return in the try block.

The new async/async* translation revealed a bug in the scoping of
labels.  Labels from enclosing functions were visible in nested
functions.  That bug is fixed here.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/536237014 .